functions/requirements.txt
# Cloud Functions Dependencies for CivicConnect
# Gen 2 Cloud Functions with Gemini AI

firebase-functions>=0.4.0
google-genai>=1.0.0
flask>=2.0.0
